JavaScript 販売手数料・販売利益 再作成
