+++ app/javascript/packs/application.js
@@ -6,7 +6,7 @@
 require("@rails/ujs").start()
 require("@rails/activestorage").start()
 require("channels")
-require("../calculation");
+require("../item_price");
 
 
 // Uncomment to copy all static images under ../images to the output folder and reference
